requirements.txt

Brotli==1.0.9
cairocffi==1.2.0
catfish==4.16.3
cbor==1.0.0
censys==2.0.9
certifi==2020.6.20
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.6
cheroot==8.5.2+ds1
CherryPy==18.6.1
cherrypy-cors==1.6
cli-helpers==2.2.1
click==8.0.3
click-plugins==1.1.1
colorama==0.4.4
commonmark==0.9.1
configobj==5.0.6
constantly==15.1.0
crackmapexec==5.2.2
crit==0.0.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
Cython==0.29.24
dataclasses-json==0.5.6
dbus-python==1.2.18
debtags==2.1
decorator==4.4.2
defusedxml==0.7.1
deprecation==2.0.7
dicttoxml==1.7.4
distro==1.6.0
distro-info==1.1
Django==3.2.10
django-extensions==3.1.5
django-sslserver==0.22
dnslib==0.9.19
dnspython==2.2.0
docopt==0.6.2
donut-shellcode==0.9.2
dropbox==11.25.0
ecdsa==0.18.0b1
email-validator==1.1.2
et-xmlfile==1.0.1
ExifRead==2.3.2
faraday-agent-parameters-types==1.0.3
faraday-client==1.1.0
faraday-plugins==1.5.9
faradaysec==3.19.0
fastapi==0.73.0
feedparser==6.0.8
fierce==1.5.0
filedepot==0.5.2
filteralchemy==0.1.0
flasgger==0.9.5
Flask==2.0.1
Flask-Classful==0.14.1
Flask-KVSession-fork==0.6.3
Flask-Limiter==1.0.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-RESTful==0.3.9
Flask-Security-Too==4.0.0
Flask-SocketIO==5.0.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
flatbuffers===1.12.1-git20200711.33e2d80-dfsg1-0.6
fonttools==4.28.5
frozenlist==1.2.0
fs==2.4.12
fuse-python==1.0.2
future==0.18.2
GDAL==3.4.1
geoip2==2.9.0
geojson==2.5.0
gitdb==4.0.9
GitPython==3.1.24
gpg===1.16.0-unknown
graphene==2.1.9
graphene-sqlalchemy==2.1.2
graphql-core==2.2.1
graphql-relay==2.0.1
greenlet==1.1.2
h11==0.12.0
h2==4.0.0
hashID==3.1.4
hiredis==1.0.1
hpack==4.0.0
html2text==2020.1.16
html5lib==1.1
httpagentparser==1.9.1
httplib2==0.20.2
humanize==0.0.0
hupper==1.10.3
hyperframe==6.0.0
hyperlink==21.0.0
icalendar==4.0.3
idna==3.2
impacket==0.9.24
importlib-metadata==4.6.4
incremental==17.5.0
iniconfig==0.0.0
invoke==1.4.1
ipaddr==2.2.0
iptools==0.6.1
ipwhois==0.15.1
IPy==1.1
ipython==7.31.1
itsdangerous==2.0.1
jaraco.classes==3.2.1
jaraco.collections==3.4.0
jaraco.functools==3.4.0
jaraco.text==3.6.0
jdcal==1.0
jedi==0.18.0
Jinja2==3.0.1
jq==1.2.1
jsonschema==3.2.0
kaitaistruct==0.9
kali-tweaks==2022.1.2
KismetCaptureBtGeiger==2021.7.1
KismetCaptureFreaklabsZigbee==2018.7.0
KismetCaptureRtl433==2020.10.1
KismetCaptureRtladsb==2020.10.1
KismetCaptureRtlamr==2020.10.1
kiwisolver==1.3.2
ldap3==2.8.1
ldapdomaindump==0.9.3
lightdm-gtk-greeter-settings==1.2.2
limiter==0.1.2
limits==1.5.1
llvmlite==0.38.0
louis==3.19.0
lsassy==3.1.1
lxml==4.7.1
lz4==3.1.3+dfsg
macholib==1.15.2
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==2.0.1
marshmallow==3.13.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
maxminddb==2.0.3
mechanize==0.4.7
minidump==0.0.21
minikerberos==0.2.14
mistune==0.8.4
mitmproxy==6.0.2
mnemonic==0.19
more-itertools==8.10.0
mpmath==0.0.0
msgpack==1.0.2
msldap==0.3.30
multidict==5.1.0
mypy-extensions==0.4.3
mysql-connector-python-rf==2.2.2
mysqlclient==1.4.6
Naked==0.1.31
nassl==4.0.0
neo4j==1.7.0.dev0
neobolt==1.7.17
neotime==1.7.4
netaddr==0.8.0
netifaces==0.11.0
networkx==2.5
nplusone==1.0.0
ntlm-auth==1.4.0
numba==0.52.0
numexpr==2.8.1
numpy==1.21.5
odfpy==1.4.2
olefile==0.46
onboard==1.4.1
openpyxl==3.0.9
orjson==0.0.0
packaging==21.3
pandas==1.3.5
paramiko==2.8.1
parso==0.8.1
passlib==1.7.4
Paste==3.5.0
PasteDeploy==2.1.1
PasteScript==2.0.2
patator==0.9
pcapy==0.11.5.dev0
pefile==2021.9.3
pendulum==2.1.2
pexpect==4.8.0
pgcli==3.3.1
pgspecial==1.11.10
phonenumbers==8.12.1
pickleshare==0.7.5
Pillow==9.0.0
pipreqs==0.4.11
plaster==1.0
plaster-pastedeploy==0.5
pluggy==0.13.0
pluginbase==1.0.1
ply==3.11
portend==3.0.0
prettytable==0.0.0
promise==2.3
prompt-toolkit==3.0.26
protobuf==3.12.4
psycopg2==2.9.2
ptyprocess==0.7.0
publicsuffix2==2.20191221
publicsuffixlist==0.7.10
py==1.10.0
py-ubjson==0.16.1
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pycares==4.1.2
pycparser==2.21
pycryptodomex==3.11.0
pycups==2.0.1
pycurl==7.44.1
pydantic==1.9.0
PyDispatcher==2.0.5
pydot==1.4.2
pyee==7.0.2
pyExploitDb==0.2.0
pyfiglet==0.8.post0
pygexf==0.2.2
Pygments==2.7.1
PyGObject==3.42.0
pygraphviz==1.7
PyHamcrest==2.0.2
pyinotify==0.9.6
PyInstaller==3.5+498e6ee058
pylibacl==0.6.0
pylnk3==0.4.2
pyminifier==2.1
pymongo==4.0.1
pymssql==2.2.2
PyMySQL==1.0.2
PyNaCl==1.5.0
PyOpenGL==3.1.5
pyOpenSSL==21.0.0
pyotp==2.3.0
pyparsing==2.4.7
PyPDF2==1.26.0
pyperclip==1.8.2
pypng==0.0.20
pyppeteer==0.2.5
pyproj==3.3.0
pypsrp==0.5.0
pypykatz==0.4.9
PyQRCode==1.2.1
PyQt5==5.15.6
PyQt5-sip==12.9.0
pyqtgraph==0.12.3
pyramid==2.0
pyrsistent==0.15.5
pyserial==3.5b0
pyShodan==0.2.3
pyshp==2.1.3
PySimpleSOAP==1.16.2
pysmbc==1.0.23
pysmi==0.3.2
pysnmp==4.4.12
PySocks==1.7.1
pyspnego==0.1.5
pytest==6.2.5
python-apt==2.3.0+b1
python-dateutil==2.8.1
python-debian==0.1.43
python-debianbts==3.2.0
python-docx==0.8.11
python-dotenv==0.19.2
python-engineio==4.0.0
python-magic==0.4.24
python-pam==1.8.4
python-pptx==0.6.18
python-snappy==0.5.3
python-socketio==5.0.3
PyTrie==0.4.0
pytz==2021.3
pytzdata==2020.1
pywerview==0.3.2
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.4.1
qrcode==7.3.1
Quamash==0.6.1
redis==3.5.3
reportbug==11.3.0
repoze.lru==0.7
requests==2.25.1
requests-file==1.5.1
requests-ntlm==1.1.0
requests-toolbelt==0.9.1
responses==0.13.4
retrying==1.3.3
rich==10.16.2
Routes==2.5.1
rq==1.10.1
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
rule-engine==1.1.0
Rx==3.2.0
scapy==2.4.4
scipy==1.7.3
secure==0.3.0
selenium==4.0.0a1
service-identity==18.1.0
setproctitle==1.2.2
sgmllib3k==1.0.0
shodan==1.26.1
simplejson==3.17.5
simplekv==0.13.0
six==1.16.0
slowapi==0.1.4
smmap==5.0.0
smoke-zephyr==2.0.1
sniffio==1.2.0
sortedcontainers==2.1.0
soupsieve==2.3.1
speaklater==1.4
spyse-python==2.2.3
SQLAlchemy==1.4.23
sqlalchemy-schemadisplay==1.3
SQLAlchemy-Utc==0.11.0
sqlparse==0.4.2
sslyze==5.0.1
starlette==0.16.0
stone==3.2.1
sympy==1.7.1
syslog-rfc5424-formatter==1.2.2
tables==3.7.0
tabulate==0.8.9
Tempita==0.5.2
tempora==4.1.2
termcolor==1.1.0
terminaltables==3.1.0
texttable==1.6.3
theHarvester==4.0.3
tld==0.11.11
tls-parser==1.2.2
token-bucket==0.3.0
toml==0.10.2
tornado==6.1
tqdm==4.57.0
traitlets==5.1.1
translationstring==1.4
Twisted==20.3.0
txaio==21.2.1
typing-extensions==3.10.0.2
typing-inspect==0.7.1
tzlocal==2.1
u-msgpack-python==2.3.0
ufoLib2==0.12.1
ujson==5.1.0
unattended-upgrades==0.1
unicodecsv==0.14.1
Unidecode==1.3.2
UNKNOWN==0.0.0
urllib3==1.26.5
urwid==2.1.2
uvicorn==0.15.0
uvloop==0.16.0
venusian==3.0.0
wafw00f==2.1.0
wapiti3==3.0.4
wcwidth==0.1.9
webargs==8.0.1
webencodings==0.5.1
WebOb==1.8.6
websocket-client==1.2.1
websockets==9.1
websockify==0.10.0
Werkzeug==2.0.2
wfuzz==3.1.0
whois==0.8
wifite==2.5.8
winacl==0.1.2
wsaccel==0.6.3
wsproto==1.0.0
WTForms==2.2.1
xcffib==0.11.1
xdg==5
xlrd==1.2.0
XlsxWriter==3.0.2
xlutils==2.0.0
xlwt==1.3.0
xmltodict==0.12.0
yara-python==4.0.4
yarg==0.1.9
yarl==1.6.3
yaswfp==0.9.3
zc.lockfile==2.0
zipp==1.0.0
zlib-wrapper==0.1.3
zope.deprecation==4.4.0
zope.interface==5.4.0
